Split test_python_types.cpp into builtin_casters, stl and pytypes
